phase 5 start

+++ vmbuild/common/Common.Phases.ps1
@@ -72,7 +72,9 @@ function Start-Phase {
     if ($deployConfig.parameters.ExistingDCName -and $Phase -eq 1) {
         Write-Log "Attempting to remove existing DNS Records"
         foreach ($item in $deployConfig.virtualMachines | Where-Object { -not ($_.hidden) } ) {
-            Remove-DnsRecord -DCName $existingDC -Domain $deployConfig.vmOptions.domainName -RecordToDelete $item.vmName
+            if ($existingDC) {
+                Remove-DnsRecord -DCName $existingDC -Domain $deployConfig.vmOptions.domainName -RecordToDelete $item.vmName
+            }
         }
     }
 
@@ -316,7 +318,8 @@ function Get-ConfigurationData {
         "2" { $cd = Get-Phase2ConfigurationData -deployConfig $deployConfig }
         "3" { $cd = Get-Phase3ConfigurationData -deployConfig $deployConfig }
         "4" { $cd = Get-Phase4ConfigurationData -deployConfig $deployConfig }
-        "5" { $cd = Get-AOandSCCMConfigurationData -deployConfig $deployConfig }
+        "5" { $cd = Get-Phase5ConfigurationData -deployConfig $deployConfig }
+        "6" { $cd = Get-AOandSCCMConfigurationData -deployConfig $deployConfig }
         Default { return }
     }
 
@@ -418,7 +421,109 @@ function Get-Phase4ConfigurationData {
 
     return $cd
 }
+function Get-Phase5ConfigurationData {
+    param (
+        [object]$deployConfig
+    )
 
+    $primaryNode = $deployConfig.virtualMachines | Where-Object { $_.role -eq "SQLAO" -and $_.OtherNode }
+    $netbiosName = $deployConfig.vmOptions.domainName.Split(".")[0]
+    $domainNameSplit = ($deployConfig.vmOptions.domainName).Split(".")
+    $dcName = $deployConfig.virtualMachines | Where-Object { $_.role -eq "DC" }
+
+
+    $NumberOfNodesAdded = 0
+    # Configuration Data
+    $cd = @{
+        AllNodes = @(
+            @{
+                NodeName = $dcName
+                Role     = 'DC'
+            }
+        )
+    }
+
+    if ($primaryNode) {
+        $SqlAgentServiceAccount = $netbiosName + "\" + $deployConfig.SQLAO.SqlAgentServiceAccount
+        $SqlServiceAccount = $netbiosName + "\" + $deployConfig.SQLAO.SqlServiceAccount
+        if (-not $primaryNode.fileServerVM) {
+            write-Log -Failure "Could not get fileServerVM name from primaryNode.fileServerVM"
+            return
+        }
+
+        $ADAccounts = @()
+        $ADAccounts += $primaryNode.vmName + "$"
+        $ADAccounts += $primaryNode.OtherNode + "$"
+        $ADAccounts += $primaryNode.ClusterName + "$"
+
+        $ADAccounts2 = @()
+        $ADAccounts2 += $($domainNameSplit[0]) + "\" + $primaryNode.vmName + "$"
+        $ADAccounts2 += $($domainNameSplit[0]) + "\" + $primaryNode.OtherNode + "$"
+        $ADAccounts2 += $($domainNameSplit[0]) + "\" + $primaryNode.ClusterName + "$"
+        $ADAccounts2 += $($domainNameSplit[0]) + "\" + $deployConfig.vmOptions.adminName
+
+        #$siteServer = $deployConfig.virtualMachines | Where-Object { $_.remoteSQLVM -eq $primaryNode.vmName }
+        #$db_name = $null
+        #if ($siteServer -and ($deployConfig.cmOptions.install)) {
+        #    $db_name = "CM_" + $siteServer.SiteCode
+        #}
+        $primary = @{
+            # Replace with the name of the actual target node.
+            NodeName        = $primaryNode.vmName
+
+            # This is used in the configuration to know which resource to compile.
+            Role            = 'ClusterNode1'
+            CheckModuleName = 'SqlServer'
+            Address         = $deployConfig.vmOptions.network
+            AddressMask     = '255.255.255.0'
+            Name            = 'Domain Network'
+            Address2        = '10.250.250.0'
+            AddressMask2    = '255.255.255.0'
+            Name2           = 'Cluster Network'
+            InstanceName    = $primaryNode.sqlInstanceName
+
+        }
+
+        $cd.AllNodes += $primary
+
+        $secondary = @{
+            # Replace with the name of the actual target node.
+            NodeName = $primaryNode.OtherNode
+
+            # This is used in the configuration to know which resource to compile.
+            Role     = 'ClusterNode2'
+        }
+        $cd.AllNodes += $secondary
+        #added Primary And Secondary
+        $NumberOfNodesAdded = $NumberOfNodesAdded + 2
+        $all = @{
+            NodeName                    = "*"
+            PSDscAllowDomainUser        = $true
+            PSDscAllowPlainTextPassword = $true
+            ClusterName                 = $primaryNode.ClusterName
+            ClusterIPAddress            = $deployConfig.SQLAO.ClusterIPAddress + "/24"
+            AGIPAddress                 = $deployConfig.SQLAO.AGIPAddress + "/255.255.255.0"
+            PrimaryReplicaServerName    = $primaryNode.vmName + "." + $deployConfig.vmOptions.DomainName
+            SecondaryReplicaServerName  = $primaryNode.OtherNode + "." + $deployConfig.vmOptions.DomainName
+            SqlAgentServiceAccount      = $SqlAgentServiceAccount
+            SqlServiceAccount           = $SqlServiceAccount
+            ClusterNameAoG              = $deployConfig.SQLAO.AlwaysOnName
+            ClusterNameAoGFQDN          = $deployConfig.SQLAO.AlwaysOnName + "." + $deployConfig.vmOptions.DomainName
+            WitnessShare                = "\\" + $primaryNode.fileServerVM + "\" + $deployConfig.SQLAO.WitnessShare
+            BackupShare                 = "\\" + $primaryNode.fileServerVM + "\" + $deployConfig.SQLAO.BackupShare
+            #Dont pass DBName, or DSC will create the database and add it to Ao.. In this new method, we install SCCM direct to AO
+            #DBName                      = $db_name
+            #ClusterIPAddress            = '10.250.250.30/24'
+        }
+        $cd.AllNodes += $all
+
+    }
+
+    if ($NumberOfNodesAdded -eq 0) {
+        return
+    }
+    return $cd
+}
 function Get-AOandSCCMConfigurationData {
     param (
         [object]$deployConfig
